+++ R/RcppExports.R
@@ -216,6 +216,18 @@ ip_in_range <- function(ip_addresses, ranges) {
     .Call('iptools_ip_in_range', PACKAGE = 'iptools', ip_addresses, ranges)
 }
 
+#'@title check if IP address falls within any of the ranges specified
+#'@description \code{ip_in_any} checks whether a vector of IP addresses
+#'fall within any of the speficied ranges.
+#'
+#'@param ip_addresses character vector of IP addresses
+#'@param ranges character vector of CIDR reanges
+#'@return a logical vector of whether a given IP was in any of the ranges
+#'@export
+ip_in_any <- function(ip_addresses, ranges) {
+    .Call('iptools_ip_in_any', PACKAGE = 'iptools', ip_addresses, ranges)
+}
+
 #'@title check whether IPv4 ranges are valid
 #'@description \code{validate_range} checks whether
 #'a vector of IPv4 CIDR ranges ("127.0.0.1/32") are valid or not.