+++ NEWS.md
@@ -4,6 +4,7 @@ iptools 0.5.0 [WIP]
 * NA support added for IP validity checkers and classifiers
 * Hilbert coordiates generator function added
 * `ip_in_any()` added for a more generic range membership test
+* `ip_to_asn()` added (with supporting functions)
 
 iptools 0.4.0
 =============